#58 構造整理のためのリファクタリング
