+++ frontend/src/components/forms/PlanDetailItem.jsx
