+++ frontend/src/components/maps/CommonInfoWindow.jsx
