+++ frontend/src/components/maps/NewSpotInfoWindow.jsx
