+++ frontend/src/components/maps/RegisteredSpotInfoWindow.jsx
