+++ frontend/src/hooks/maps/usePlaceSelection.js
