+++ frontend/src/hooks/trips/usePlanOverview.js
@@ -1,5 +1,5 @@
 import { useState, useEffect, useRef, useMemo } from "react";
-import { fetchPlanOverviewData, planOverviewUpdate } from "@/api/planOverviewApi";
+import { fetchPlanOverviewData, planOverviewUpdate } from "@/services/api/planOverviewApi";
 import { parseError, ERROR_MESSAGES } from '@/utils/errorHandler';
 
 export const usePlanOverview = (planId) => {
