+++ frontend/src/main.jsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
-import { Provider } from "./components/ui/provider"
+import { Provider } from "@/components/ui/provider"
 import { createRoot } from 'react-dom/client'
 import { RouterProvider } from 'react-router-dom'
 
-import { routesBasic } from './routes/routesBasic.jsx'
-import { AuthProvider } from './contexts/AuthContext'
+import { routesBasic } from '@/routesBasic.jsx'
+import { AuthProvider } from '@/store/AuthContext'
 // import './styles/index.css'
 
 createRoot(document.getElementById('root')).render(
