+++ frontend/src/services/api/planDetailApi.js
@@ -1,4 +1,4 @@
-import { api } from "@/api/api";
+import { api } from "./api";
 import { formatPlanDetail } from "@/utils/planDataFormatter";
