+++ frontend/src/services/api/planOverviewApi.js
@@ -1,4 +1,4 @@
-import { api } from "@/api/api";
+import { api } from "./api";
 import { formatPlanOverview } from "@/utils/planDataFormatter";
 
 // 旅行概要の取得
