+++ frontend/src/store/AuthContext.jsx
@@ -1,6 +1,6 @@
 // src/contexts/AuthContext.jsx
 import { createContext, useContext, useState, useEffect } from 'react';
-import { api } from '@/api/api';
+import { api } from '@/services/api/api';
 
 const AuthContext = createContext();
