+++ frontend/src/utils/validation/schemas.jsx
@@ -1,0 +1,9 @@
+import { tripSchema } from "@/utils/validation/tripSchema";
+import { planDetailSchema } from "@/utils/validation/planDetailSchema";
+
+//バリデーションスキーマをまとめて定義
+
+export const schemas = {
+  tripSchema,
+  planDetailSchema,
+};
