--- frontend/src/validation/index.jsx
@@ -1,7 +1,0 @@
-import { tripSchema } from "./tripSchema";
-import { planDetailSchema } from "./planDetailSchema";
-
-export const schemas = {
-  tripSchema,
-  planDetailSchema,
-};
